+++ satellite/src/device-types/loupedeck.ts
