+++ satellite/src/surface-manager.ts
@@ -7,7 +7,7 @@ import { QuickKeysPlugin } from './device-types/xencelabs-quick-keys.js'
 import * as HID from 'node-hid'
 import { wrapAsync } from './lib.js'
 import { InfinittonPlugin } from './device-types/infinitton.js'
-import { LoupedeckPlugin } from './device-types/loupedeck-plugin.js'
+import { LoupedeckPlugin } from './device-types/loupedeck.js'
 import { ApiSurfaceInfo, ApiSurfacePluginInfo, ApiSurfacePluginsEnabled } from './apiTypes.js'
 import { BlackmagicControllerPlugin } from './device-types/blackmagic-panel.js'
 import { SurfaceProxy, SurfaceProxyContext } from './surfaceProxy.js'
